refactor(lotto): 패키지 이동
